requirements.txt
certifi==2025.11.12
chardet==4.0.0
charset-normalizer==3.4.4
colorama==0.4.6
elasticsearch==7.10.1
elasticsearch-dsl==7.3.0
filelock==3.19.1
fsspec==2025.9.0
huggingface-hub==0.36.0
idna==2.10
Jinja2==3.1.6
MarkupSafe==2.1.5
modelscope==1.32.0
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==1.9.2
pygame==2.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.25.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.1
typing_extensions==4.15.0
urllib3==1.26.20
wheel==0.45.1
